Move models to a file
I'm not sure why this has always been in a method. Perhaps there was a
more complicated teardown step after the tests ran...

I added the following to the `setup` in `test_arhp_wrong_db.rb`:

```
ActiveRecordHostPool::PoolProxy.class_variable_set(:@@_connection_pools, {})
Phenix.burn!
```

Rails checks that existance of the db configs when you call `connects_to` on a
model meaning we need to call `Phenix.rise!` (thereby loading the db
configs) _before_ we load the models. This broke the
`test_arhp_wrong_db` tests because they test ARHP's behavior when a
database doesn't exist. So now, for those tests, we drop all of the dbs
before they run.
