+++ test/test_arhp.rb
@@ -6,7 +6,6 @@ class ActiveRecordHostPoolTest < Minitest::Test
   include ARHPTestSetup
   def setup
     Phenix.rise! config_path: "test/three_tier_database.yml"
-    arhp_create_models
   end
 
   def teardown
